+++ src/main/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/AdminAmendLocationDetailsAboutToStart.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscs.ccd.presubmit.adminamendlocationdetails;
+package uk.gov.hmcts.reform.sscs.ccd.presubmit;
 
 import static java.util.Objects.requireNonNull;
 
@@ -13,7 +13,6 @@ import uk.gov.hmcts.reform.sscs.ccd.callback.Callback;
 import uk.gov.hmcts.reform.sscs.ccd.callback.CallbackType;
 import uk.gov.hmcts.reform.sscs.ccd.callback.PreSubmitCallbackResponse;
 import uk.gov.hmcts.reform.sscs.ccd.domain.*;
-import uk.gov.hmcts.reform.sscs.ccd.presubmit.PreSubmitCallbackHandler;
 import uk.gov.hmcts.reform.sscs.service.RegionalProcessingCenterService;
 import uk.gov.hmcts.reform.sscs.service.VenueService;
