+++ src/test/java/uk/gov/hmcts/reform/sscs/ccd/presubmit/AdminAmendLocationDetailsAboutToStartTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.sscs.ccd.presubmit.adminamendlocationdetails;
+package uk.gov.hmcts.reform.sscs.ccd.presubmit;
 
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
